requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
qrcode[pil]==8.2
pillow==11.2.1